requirements.txt
selenium
lxml
asgiref
Django
djangorestframework
drf-yasg
inflection
# mysqlclient
packaging
pytz
PyYAML
sqlparse
typing_extensions
tzdata
uritemplate
drf-api-logger
django-environ
beautifulsoup4
mysql-connector-python
APScheduler
Pillow
firebase-admin
djangorestframework-simplejwt
django-cors-headers
# kobart
# git+https://github.com/SKT-AI/KoBART#egg=kobart
torch
transformers
pytorch-lightning